+++ packages/twenty-server/src/engine/metadata-modules/field-metadata/constants/field-metadata-standard-overrides-properties.constant.ts
@@ -1,0 +1,7 @@
+import { FlatFieldMetadataPropertiesToCompare } from "src/engine/metadata-modules/flat-field-metadata/types/flat-field-metadata-properties-to-compare.type";
+
+export const FIELD_METADATA_STANDARD_OVERRIDES_PROPERTIES = [
+  'label',
+  'description',
+  'icon',
+] as const satisfies FlatFieldMetadataPropertiesToCompare[];
