+++ packages/twenty-server/src/engine/metadata-modules/field-metadata/types/field-metadata-standard-overrides-properties.type.ts
@@ -1,0 +1,4 @@
+import { FIELD_METADATA_STANDARD_OVERRIDES_PROPERTIES } from "src/engine/metadata-modules/field-metadata/constants/field-metadata-standard-overrides-properties.constant";
+
+export type FieldMetadataStandardOverridesProperties =
+  (typeof FIELD_METADATA_STANDARD_OVERRIDES_PROPERTIES)[number];
