+++ packages/twenty-server/src/engine/metadata-modules/flat-object-metadata/utils/from-update-object-input-to-flat-object-metadata.util.ts
@@ -9,15 +9,13 @@ import { findFlatObjectMetadataInFlatObjectMetadataMaps } from 'src/engine/metad
 import { FLAT_OBJECT_METADATA_PROPERTIES_TO_COMPARE } from 'src/engine/metadata-modules/flat-object-metadata/constants/flat-object-metadata-properties-to-compare.constant';
 import { FlatObjectMetadataPropertiesToCompare } from 'src/engine/metadata-modules/flat-object-metadata/types/flat-object-metadata-properties-to-compare.type';
 import { type FlatObjectMetadata } from 'src/engine/metadata-modules/flat-object-metadata/types/flat-object-metadata.type';
-import {
-  type ObjectMetadataStandardOverridesProperties,
-  objectMetadataStandardOverridesProperties,
-} from 'src/engine/metadata-modules/object-metadata/dtos/object-standard-overrides.dto';
+import { OBJECT_METADATA_STANDARD_OVERRIDES_PROPERTIES } from 'src/engine/metadata-modules/object-metadata/constants/object-metadata-standard-overrides-properties.constant';
 import { type UpdateOneObjectInput } from 'src/engine/metadata-modules/object-metadata/dtos/update-object.input';
 import {
   ObjectMetadataException,
   ObjectMetadataExceptionCode,
 } from 'src/engine/metadata-modules/object-metadata/object-metadata.exception';
+import { ObjectMetadataStandardOverridesProperties } from 'src/engine/metadata-modules/object-metadata/types/object-metadata-standard-overrides-properties.types';
 import { isStandardMetadata } from 'src/engine/metadata-modules/utils/is-standard-metadata.util';
 
 type FromUpdateObjectInputToFlatObjectMetadataArgs = {
@@ -67,7 +65,7 @@ export const fromUpdateObjectInputToFlatObjectMetadata = ({
       updatedEditableObjectProperties,
     ).filter(
       (property) =>
-        !objectMetadataStandardOverridesProperties.includes(
+        !OBJECT_METADATA_STANDARD_OVERRIDES_PROPERTIES.includes(
           property as ObjectMetadataStandardOverridesProperties,
         ),
     );
@@ -80,7 +78,7 @@ export const fromUpdateObjectInputToFlatObjectMetadata = ({
     }
 
     const updatedStandardFlatObjectdMetadata =
-      objectMetadataStandardOverridesProperties.reduce((acc, property) => {
+      OBJECT_METADATA_STANDARD_OVERRIDES_PROPERTIES.reduce((acc, property) => {
         const isPropertyUpdated =
           updatedEditableObjectProperties[property] !== undefined;
