+++ packages/twenty-server/src/engine/metadata-modules/object-metadata/constants/object-metadata-standard-overrides-properties.constant.ts
@@ -1,0 +1,8 @@
+import { FlatObjectMetadataPropertiesToCompare } from "src/engine/metadata-modules/flat-object-metadata/types/flat-object-metadata-properties-to-compare.type";
+
+export const OBJECT_METADATA_STANDARD_OVERRIDES_PROPERTIES = [
+  'labelSingular',
+  'labelPlural',
+  'description',
+  'icon',
+] as const satisfies FlatObjectMetadataPropertiesToCompare[];
