+++ packages/twenty-server/src/engine/metadata-modules/object-metadata/types/object-metadata-standard-overrides-properties.types.ts
@@ -1,0 +1,4 @@
+import { OBJECT_METADATA_STANDARD_OVERRIDES_PROPERTIES } from "src/engine/metadata-modules/object-metadata/constants/object-metadata-standard-overrides-properties.constant";
+
+export type ObjectMetadataStandardOverridesProperties =
+  (typeof OBJECT_METADATA_STANDARD_OVERRIDES_PROPERTIES)[number];
